feat: Adicionado codigo para impressão raw em sistemas Unix, necessita de testes.

+++ src/OpenAC.Net.Devices/Commom/ByteArrayBuilder.cs
@@ -8,7 +8,7 @@
 // ***********************************************************************
 // <copyright file="ByteArrayBuilder.cs" company="OpenAC .Net">
 //		        		   The MIT License (MIT)
-//	     		    Copyright (c) 2016 Projeto OpenAC .Net
+//	     		    Copyright (c) 2014 - 2024 Projeto OpenAC .Net
 //
 //	 Permission is hereby granted, free of charge, to any person obtaining
 // a copy of this software and associated documentation files (the "Software"),
@@ -154,7 +154,7 @@ namespace OpenAC.Net.Devices.Commom
         /// <param name="position">the position to seek.</param>
         public void Seek(int position)
         {
-            store.Seek((long)position, SeekOrigin.Begin);
+            store.Seek(position, SeekOrigin.Begin);
         }
 
         /// <summary>
